feat(pipette): Read pipette information from serial (#437)
* feat(pipettes): read name from serial

The serial number now will be sent from the host as a pre-packed mixed
binary and ascii message.

The pipette names are now encoded as binary uint16s like the models. The
datecode is unchanged.

The name is now actually parsed from the serial number in response to
GetPipetteInfo.

Closes RET-1169

* refactor(simulators): add command line options

We had a bunch of options we could pass to the simulators but they were
documented in source and docs and not in an inspectable way. It also
makes it annoying to add more.

Instead, or really in addition, we can use boost::program_options to
parse the environment variables. We can also really easily use it to
provide command line flags alongside them, and then we get a -h option
that can document both.

While we're at it, add a new option for the eeprom that specifies a file
to back up the eeprom contents.

* refactor(eeprom): use a file to back up the eeprom

This lets eeprom changes be persistent between runs, if needed. We can
take advantage of the new options to specify a filename, and use normal
file handling utilities (not fstreams because they are slightly
nightmarish to deal with) to handle and write to files.

Also, by default the pipette simulator now uses a file backing for its
eeprom when you run with ./run_simulators.

* refactor(sensors): smaller data in cap sensor sim

The simulator harness for the capacitive sensor returned numbers large
enough that the capdac would be adjusted, but then the harness didn't
take the altered capdac into account. That means the numbers returned on
can would change over time. By reducing the values returned from the
sensor, we fix the problem.


* don't use append to open eeprom files

It makes the beginning of the file the end of what was there, you can't
overwrite. We have to try to open for r+ first, which won't overwrite
anything but will fail if the file doesn't exist, and then try w+ if
that fails

* allow full space for serial number datecodes

* format cap sense handler

* pipette: enable gpio ports more reliably

+++ can/tests/test_messages.cpp
@@ -228,7 +228,7 @@ SCENARIO("message serializing works") {
                 REQUIRE(len == sizeof(data));
                 // check that the data has been serialized till the end of the
                 // buffer
-                for (uint i = 0; i < data.size() && iter < arr.end(); i++) {
+                for (size_t i = 0; i < data.size() && iter < arr.end(); i++) {
                     REQUIRE(*(iter++) == data[i]);
                 }
             }
@@ -253,11 +253,11 @@ SCENARIO("message serializing works") {
                 // check length was serialized correctly
                 REQUIRE(len == sizeof(data));
                 // check that the data has been serialized
-                for (uint i = 0; i < data.size() && iter < arr.end(); i++) {
+                for (size_t i = 0; i < data.size() && iter < arr.end(); i++) {
                     REQUIRE(*(iter++) == data[i]);
                 }
                 // check that the remainder of the buffer is 0x00
-                for (uint i = data.size();
+                for (size_t i = data.size();
                      i < eeprom::types::max_data_length && iter < arr.end();
                      i++) {
                     REQUIRE(*(iter++) == 0x00);
