+++ eeprom/tests/test_revision.cpp
@@ -28,7 +28,7 @@ SCENARIO("Writing revision") {
                          0));
                 message::WriteEepromMessage write_message;
                 types::data_length expected_bytes;
-                for (ulong i = 0; i < queue_client.messages.size(); i++) {
+                for (size_t i = 0; i < queue_client.messages.size(); i++) {
                     expected_bytes =
                         ((addresses::revision_length -
                           (i * types::max_data_length)) > types::max_data_length
@@ -73,7 +73,7 @@ SCENARIO("Reading revision") {
 
                 types::data_length expected_bytes;
                 message::ReadEepromMessage read_message;
-                for (ulong i = 0; i < queue_client.messages.size(); i++) {
+                for (size_t i = 0; i < queue_client.messages.size(); i++) {
                     expected_bytes =
                         ((addresses::revision_length -
                           (i * types::max_data_length)) > types::max_data_length
